Worker: Improve queue handling for initial configuration request
The `exclusive` queue is problematic. For example, with GarageMQ, the
Worker would not properly consume AMQP messages after stopping it once
with SIGTERM and starting it again.

The reason is that it would blocking-wait for the exclusivity to be
released, which is apparently still held as a lock on the previous
connection to the AMQP broker, which is now lost.

The hope is that the hard exclusivity is not needed here, but instead,
a random queue name will provide a kind of soft-exclusivity.