Data model for responses from Actions, instead of using the filesystem
Channeling responses from Actions through Zip files was error prone and
also induces drag on the filesystem.

Now, with Python 3's new `ThreadPoolExecutor`, it is easy to defer
functions to a thread, but still get their return values back, by using
futures.

The return value from an Action's `execute()` method must now be an
instance of `ActionResponse`, which is a container object for any number
of `FileResponse` objects and friends.