Notifier: Modernize `notify` interface
The main `notify()` method now accepts a `NotificationMessage` object
instance instead of a dictionary. This is both more convenient and
extendable.

Channeling responses from Actions through the filesystem is tedious and
error prone. Now it is done in-memory, so all notifiers will not have
to go to the filesystem to receive the corresponding information.

Now, it will be right there within the `NotificationMessage` object, at
your fingertips.